requirements.txt
aniso8601==8.1.0
astroid==2.4.2
beautifulsoup4==4.9.3
bidict==0.19.0
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dataclasses==0.6
decorator==4.4.2
diskcache==5.1.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
Flask-Uploads==0.2.1
gunicorn==20.0.4
httpie==2.3.0
idna==2.10
isodate==0.6.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsobject==0.10.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.0.0b11
mccabe==0.6.1
networkx==2.2
numpy==1.19.4
obonet==0.2.6
ontobio==2.1.2
pandas==1.2.0
prefixcommons==0.1.9
Pygments==2.7.3
pylint==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
pysolr==3.9.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.3.1
rdflib==5.0.0
requests==2.25.1
requests-toolbelt==0.9.1
scipy==1.6.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.1
SPARQLWrapper==1.8.5
toml==0.10.2
urllib3==1.26.2
Werkzeug==0.16.0
wrapt==1.12.1
yamldown==0.1.8
